requirements.txt
discord.py>=2.3.2
jishaku>=2.6.0
aiohttp>=3.9.0
import-expression>=0.3.0
asyncpg>=0.29.0
python-dotenv>=1.0.0